Moving all search methods to a test class where they belong, so we don't accidentally use them in production code.
We can't consider these as production-grade search methods, since they only return the first page of results. The only exception to this is the search in getLatestCopyForDataset, which explicitly restricts the page size to 1.

Handily, this also greatly reduces the size of our client class so we can *almost* remove the number.of.methods Scalastyle suppression.