Add endpoint for more efficient deleting (#96)
* Add endpoint for more efficient deleting

With this commit we add a new endpoint in Spandex-http for deleting all of the documents (dataset
copies, columns, and field values) associated with a particular dataset. The new endpoint returns
the counts of the different types deleted from the index. It uses the existing queries for finding
all documents by dataset ID and deleting (using ES's scan-and-scroll API).

Unit tests pass. Tested locally.

* Add logging to deletion script per code review suggestion

Reviewed-by: ayn.leslie