Do some cleanup (#106)
* Do some cleanup

The primary motivation for this round of cleanup was to make the mapping handling a little more
sane before adding the new type to the index. This includes the addition of the mapping JSON files
to the resources directory, the addition of the settings JSON file to the resources directory, and
the removal of those elements from reference.conf, which was brittle and odd since those things
never change. These changes are the primary source of changes here. Additionally, I did a round of
cleanup on the README, although this will change more in upcoming commits.

* Fix failing assembly build and style errors

* Fix test compilation error

* Fix broken tests

Not worrying about the Cloudbees build error for the moment.

Reviewed-by: ayn.leslie