Upgrade Elasticsearch to 5.4.x (#113)
* EN-13673: Upgrade spandex to use ES 5.4.1

* Do some cleanup (#106)

* Do some cleanup

The primary motivation for this round of cleanup was to make the mapping handling a little more
sane before adding the new type to the index. This includes the addition of the mapping JSON files
to the resources directory, the addition of the settings JSON file to the resources directory, and
the removal of those elements from reference.conf, which was brittle and odd since those things
never change. These changes are the primary source of changes here. Additionally, I did a round of
cleanup on the README, although this will change more in upcoming commits.

* Fix failing assembly build and style errors

* Fix test compilation error

* Fix broken tests

Not worrying about the Cloudbees build error for the moment.

* Add scripts for setting up a new Spandex index (#107)

This adds two new scripts to the Spandex bin dir for setting up a new index:

- setup_index
- create_index_alias

The first mirrors the `bootstrap_catalog` script (which should be renamed). The functionality for
creating an alias is separate since it often not the case that you want to create an alias at the
time you initially create an index.

There were a few style changes in my repo that I decided to commit for consistency w/ our othr
projects.

Tests pass.

* EN-17472 Ensure suggestions are unique and use edge ngrams (#109)

* EN-17472 Ensure suggestions are unique

This fundamentally changes the way we serve suggestions. Rather than using the in-memory completion
suggester which no longer returns unique suggestions, we use an autocomplete subfield consisting of
edge ngrams that are created at index-time. This has the advantage of no longer pinning the
suggestion finite state transducer data structure to memory, with a corresponding trade-off in
performance. (Details on the impact on performance forthcoming.)

The simplifies the codebase and normalizes it with an eye towards consistency with Cetera. A
smattering of changes for consistency:

- moved ES settings and mappings out of reference configuration file into JSON resource files
- added scripts for setting up index and creating index alias

Lots of cruft was removed as a result of removing the no longer necessary CompletionAnalyzer. All
analysis is defined directly in the ES settings file.

There is a subtle change in scoring. When no text is provided to the endpoint, rather than
returning a meaningless score, we return the frequency of the values. When text is provided, the
score is the vanilla relevance score returned by Elasticsearch.

In order to return unique suggestion results, we execute a terms aggregation along with the
query (dataset, column, and optional text constraints). When query text is provided, we pass along
a max subaggregation that gets the original scores from the returned documents. Using a script to
get the max score breaks unit tests due to the fact that Elasticsearch no longer supports local ES
nodes for unit testing. Future work: break out functional/integration tests into a separate
submodule that uses the official ES integration testing framework.

Eyeballed results locally on staging corpus. Unit tests pass.

* Remove cruft from secondary watcher default conf file

* Remove references to fuzziness from default config

* EN-17476 Improve logging (#110)

This makes the request logging in Spandex mirror that of Cetera.

* Do not hardcode env in spandex log scraper (#111)

* EN-13673: Upgrade spandex to use ES 5.4.1

* Do some cleanup (#106)

* Do some cleanup

The primary motivation for this round of cleanup was to make the mapping handling a little more
sane before adding the new type to the index. This includes the addition of the mapping JSON files
to the resources directory, the addition of the settings JSON file to the resources directory, and
the removal of those elements from reference.conf, which was brittle and odd since those things
never change. These changes are the primary source of changes here. Additionally, I did a round of
cleanup on the README, although this will change more in upcoming commits.

* Fix failing assembly build and style errors

* Fix test compilation error

* Fix broken tests

Not worrying about the Cloudbees build error for the moment.

* Add scripts for setting up a new Spandex index (#107)

This adds two new scripts to the Spandex bin dir for setting up a new index:

- setup_index
- create_index_alias

The first mirrors the `bootstrap_catalog` script (which should be renamed). The functionality for
creating an alias is separate since it often not the case that you want to create an alias at the
time you initially create an index.

There were a few style changes in my repo that I decided to commit for consistency w/ our othr
projects.

Tests pass.

* EN-17472 Ensure suggestions are unique

This fundamentally changes the way we serve suggestions. Rather than using the in-memory completion
suggester which no longer returns unique suggestions, we use an autocomplete subfield consisting of
edge ngrams that are created at index-time. This has the advantage of no longer pinning the
suggestion finite state transducer data structure to memory, with a corresponding trade-off in
performance. (Details on the impact on performance forthcoming.)

The simplifies the codebase and normalizes it with an eye towards consistency with Cetera. A
smattering of changes for consistency:

- moved ES settings and mappings out of reference configuration file into JSON resource files
- added scripts for setting up index and creating index alias

Lots of cruft was removed as a result of removing the no longer necessary CompletionAnalyzer. All
analysis is defined directly in the ES settings file.

There is a subtle change in scoring. When no text is provided to the endpoint, rather than
returning a meaningless score, we return the frequency of the values. When text is provided, the
score is the vanilla relevance score returned by Elasticsearch.

In order to return unique suggestion results, we execute a terms aggregation along with the
query (dataset, column, and optional text constraints). When query text is provided, we pass along
a max subaggregation that gets the original scores from the returned documents. Using a script to
get the max score breaks unit tests due to the fact that Elasticsearch no longer supports local ES
nodes for unit testing. Future work: break out functional/integration tests into a separate
submodule that uses the official ES integration testing framework.

Eyeballed results locally on staging corpus. Unit tests pass.

* Remove cruft from secondary watcher default conf file

* Remove references to fuzziness from default config

* EN-17476 Improve logging (#110)

This makes the request logging in Spandex mirror that of Cetera.

* Address PR comments

* Cleanup (#112)

* Some cleanup that was accidentally left out of last merge

* Remove perf submodule

* Fix shard configuration value in fedramp cluster override settings

Paired-with: ayn.leslie