Log timings for resyncing docs (#115)
* Log timings for resyncing docs

This adds INFO-level logging to indicate the end-to-end time (in milliseconds) taken to resync a
document. Additionally, it adds a refresh (boolean) parameter to all of the `deleteBy...` methods
in the Spandex ES client. The reason: historically, we have forced a refresh by default and this is
costly and generally not necessary. With the added parameterization, we can explicitly call
`client.refresh()` where appropriate. I've updated cases where we were unnecessarily forcing a
refresh.

https://www.elastic.co/guide/en/elasticsearch/reference/5.4/docs-refresh.html

* Use consistent API for deleteColumnMap method (add refresh parameter)

Reviewed-by: ayn.leslie