Ensure we wait on certain events
With the recent change to refresh less aggressively, we introduced a regression. There are cases
where we assert that a change *must* be reflected in the index before proceeding. In those cases,
we have two choices: remove the assertion, or wait until the change is reflected in the index
before proceeding. The latter is acceptable in certain cases (which is reflected in this initial
change).