Fix SBT release (#143)
There was some fallout of the the recent Cloudbees changes -- specifically the SBT release task stopped working. The sbt release command was resulting in a runtime exception (java.lang.RuntimeException: Repository for publishing is not specified.). This changeset consists of a small change to the root project's build definition to address that issue.

Reviewed-by: parker.coleman